add check

+++ R/reportGenerator.R
@@ -86,7 +86,7 @@ reportGenerator <- function(logger = NULL) {
                    )
                  ),
                  fluidRow(
-                   column(width = 3,
+                   column(width = 4,
                           h2("Create report application"),
                           tags$head(tags$style(".dlAppBtn{ margin-left:15px;margin-right:15px; margin-top:25px; color:#444 !important; }
                                                 .dlCheck { margin-left:15px;margin-right:15px; margin-top:30px; color:#444 !important; }")),
@@ -523,50 +523,53 @@ reportGenerator <- function(logger = NULL) {
     })
 
     observeEvent(input$createReportApp, {
-      # Specify source directory
-      reportAppDir <- "reportApp"
-      packagePath <- system.file(reportAppDir, package = "ReportGenerator")
+      log4r::info(logger, paste("Create shiny application, reportingEnabled = ", input$enableReporting))
 
-      # Copy files
-      targetPath <- getwd()
-      file.copy(packagePath, targetPath, recursive = T)
-      reportAppPath <- file.path(targetPath, reportAppDir)
-      reportAppConfigPath <- file.path(reportAppPath, "config")
-      dir.create(reportAppConfigPath)
-
-      # Copy template
-      if (input$enableReporting) {
-        reportTemplateFile <- "DARWIN_EU_Study_Report.docx"
-        reportTemplate <- system.file("templates",
-                                      "word",
-                                      reportTemplateFile,
-                                      package = "ReportGenerator")
-        file.copy(reportTemplate, file.path(reportAppConfigPath, reportTemplateFile))
-        menuConfigFile <- "menuConfig.yaml"
-        menuConfig <- system.file("config", menuConfigFile, package = "ReportGenerator")
-        file.copy(menuConfig, file.path(reportAppConfigPath, menuConfigFile))
-
-        # update ui
-        file.remove(file.path(reportAppPath, "ui.R"))
-        file.rename(file.path(reportAppPath, "ui_total.R"), file.path(reportAppPath, "ui.R"))
-      } else {
-        file.remove(file.path(reportAppPath, "ui_total.R"))
-      }
-      # Create results dir
-      targetPathResults <- file.path(reportAppPath, "results")
-      dir.create(targetPathResults)
-
-      # Insert results from app
       uploadedFilesList <- reactiveValuesToList(uploadedFiles)
-      if (length(uploadedFilesList) > 0) {
+      uploadedFilesValues <- unlist(lapply(names(uploadedFilesList), FUN = function(name) {uploadedFilesList[[name]]}))
+      if (!is.null(uploadedFilesValues)) {
+        # Specify source directory
+        reportAppDir <- "reportApp"
+        packagePath <- system.file(reportAppDir, package = "ReportGenerator")
+
+        # Copy files
+        targetPath <- getwd()
+        file.copy(packagePath, targetPath, recursive = T)
+        reportAppPath <- file.path(targetPath, reportAppDir)
+        reportAppConfigPath <- file.path(reportAppPath, "config")
+        dir.create(reportAppConfigPath)
+
+        # Copy template
+        if (input$enableReporting) {
+          reportTemplateFile <- "DARWIN_EU_Study_Report.docx"
+          reportTemplate <- system.file("templates",
+                                        "word",
+                                        reportTemplateFile,
+                                        package = "ReportGenerator")
+          file.copy(reportTemplate, file.path(reportAppConfigPath, reportTemplateFile))
+          menuConfigFile <- "menuConfig.yaml"
+          menuConfig <- system.file("config", menuConfigFile, package = "ReportGenerator")
+          file.copy(menuConfig, file.path(reportAppConfigPath, menuConfigFile))
+
+          # update ui
+          file.remove(file.path(reportAppPath, "ui.R"))
+          file.rename(file.path(reportAppPath, "ui_total.R"), file.path(reportAppPath, "ui.R"))
+        } else {
+          file.remove(file.path(reportAppPath, "ui_total.R"))
+        }
+        # Create results dir
+        targetPathResults <- file.path(reportAppPath, "results")
+        dir.create(targetPathResults)
+
+        # Insert results from app
+        log4r::info(logger, "Add uploadedFiles and session results")
         saveRDS(list("uploadedFiles" = uploadedFilesList),
                 file.path(targetPathResults, "uploadedFiles.rds"))
-      }
 
-      itemsListObjects <- reactiveValuesToList(do.call(reactiveValues, itemsList$objects))
-      dataReportObjects <- reactiveValuesToList(do.call(reactiveValues, dataReport$objects))
-      if (length(itemsListObjects) > 0) {
-        if (length(dataReportObjects) > 0) {
+        # session files
+        itemsListObjects <- reactiveValuesToList(do.call(reactiveValues, itemsList$objects))
+        if (!is.null(dataReport$objects)) {
+          dataReportObjects <- reactiveValuesToList(do.call(reactiveValues, dataReport$objects))
           saveRDS(list("itemsList" = itemsListObjects,
                        "reportItems" = dataReportObjects),
                   file.path(targetPathResults, "session.rds"))
@@ -575,8 +578,13 @@ reportGenerator <- function(logger = NULL) {
                        "reportItems" = NULL),
                   file.path(targetPathResults, "session.rds"))
         }
+
+        log4r::info(logger, glue::glue("Shiny app has been created in {targetPath}"))
+        shinyjs::html("createAppOutput", glue::glue("<br>Shiny app created in {targetPath}"), add = TRUE)
+      } else {
+        log4r::info(logger, "No files have been uploaded, app not created.")
+        shinyjs::html("createAppOutput", glue::glue("<br>Please upload files before generating the app"), add = TRUE)
       }
-      shinyjs::html("createAppOutput", glue::glue("<br>Shiny app created in {targetPath}"), add = TRUE)
     })
 
   }
